# cali_lights_openapi.yaml
openapi: 3.0.3
info:
  title: Cali Lights API
  version: "1.0.0"
  description: >
    Networked Triad Mission platform: chains, invites, missions, gallery, metadata analysis,
    Veo video generation, and invite/bridge flows. Every endpoint maps directly to checklist B.
servers:
  - url: https://api.cali-lights.example
...
security:
  - bearerAuth: []
paths:
  /api/chain/create:
    post:
      tags: [Chains]
      summary: Create a new chain
      operationId: createChain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 1 }
                color_theme:
                  type: object
                  additionalProperties: true
                  description: Arbitrary color + motion tokens used in Orbit map styling.
              required: [name]
      responses:
        "200":
          description: Chain created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chain'
        "401": { $ref: '#/components/responses/Unauthorized' }
        "422": { $ref: '#/components/responses/UnprocessableEntity' }
...
  /api/invite/accept:
    post:
      tags: [Invites]
      summary: Accept an invite token and join a chain
      operationId: acceptInvite
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token: { type: string }
                displayName: { type: string }
                avatarUrl: { type: string, format: uri, nullable: true }
              required: [token]
      responses:
        "200":
          description: Membership created
          content:
            application/json:
              schema:
                type: object
                properties:
                  membership: { $ref: '#/components/schemas/Membership' }
                  chain: { $ref: '#/components/schemas/Chain' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "404": { $ref: '#/components/responses/NotFound' }
...
  /api/gallery/media:
    get:
      tags: [Gallery]
      summary: Paginated media entries with filters
      operationId: listMedia
      parameters:
        - $ref: '#/components/parameters/ScopeParam'
        - $ref: '#/components/parameters/ChainIdParam'
        - $ref: '#/components/parameters/UserIdParam'
        - in: query
          name: missionId
          schema: { type: string, format: uuid }
        - in: query
          name: prompt
          schema: { type: string }
        - in: query
          name: paletteBucket
          schema:
            type: string
            description: 12-step hue bucket label (e.g., warm, cool, dusk).
        - in: query
          name: tags
          schema: { type: string, description: "Comma-separated scene/object tags" }
        - in: query
          name: favoriteOnly
          schema: { type: boolean, default: false }
        - in: query
          name: sort
          schema:
            type: string
            enum: [newest, oldest, warm_to_cool, cool_to_warm, sync]
            default: newest
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        "200":
          description: Paginated media entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Entry' }
                  page: { type: integer }
                  pageSize: { type: integer }
                  total: { type: integer }
        "401": { $ref: '#/components/responses/Unauthorized' }
...
  /api/entry/commit:
    post:
      tags: [Entries]
      summary: Submit media for a mission and trigger metadata analysis
      operationId: commitEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                missionId: { type: string, format: uuid }
                mediaUrl: { type: string, format: uri }
                mediaType: { type: string, enum: [photo, video] }
                gpsCity: { type: string }
                capturedAt: { type: string, format: date-time, nullable: true }
                altText: { type: string, nullable: true }
              required: [missionId, mediaUrl, mediaType]
      responses:
        "200":
          description: Entry stored
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Entry' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "409": { $ref: '#/components/responses/Conflict' }
...
  /api/chapter/share:
    post:
      tags: [Chapters]
      summary: Toggle chapter shareability and issue signed URL
      operationId: shareChapter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                chapterId: { type: string, format: uuid }
                on: { type: boolean }
                expiresInHours: { type: integer, minimum: 1, maximum: 720, default: 72 }
              required: [chapterId, on]
      responses:
        "200":
          description: Share state
          content:
            application/json:
              schema:
                type: object
                properties:
                  chapterId: { type: string, format: uuid }
                  shareable: { type: boolean }
                  shareUrl: { type: string, format: uri, nullable: true }
                  expiresAt: { type: string, format: date-time, nullable: true }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ScopeParam:
      name: scope
      in: query
      schema: { type: string, enum: [chain, network, user], default: chain }
    ChainIdParam:
      name: chainId
      in: query
      schema: { type: string, format: uuid }
    UserIdParam:
      name: userId
      in: query
      schema: { type: string, format: uuid }
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSizeParam:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 24 }
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: Conflict with current state
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
...
    Chain:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        createdBy: { type: string, format: uuid }
        colorTheme:
          type: object
          additionalProperties: true
        activeMissionId: { type: string, format: uuid, nullable: true }
        streakDays: { type: integer, minimum: 0 }
        isPublic: { type: boolean, default: false }
        dominantHue: { type: integer, minimum: 0, maximum: 359, nullable: true }
    Membership:
      type: object
      properties:
        userId: { type: string, format: uuid }
        chainId: { type: string, format: uuid }
        role: { type: string, enum: [admin, member] }
        joinedAt: { type: string, format: date-time }
...
    Entry:
      type: object
      properties:
        id: { type: string, format: uuid }
        missionId: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        mediaUrl: { type: string, format: uri }
        mediaType: { type: string, enum: [photo, video] }
        gpsCity: { type: string, nullable: true }
        capturedAt: { type: string, format: date-time, nullable: true }
        dominantHue: { type: integer, minimum: 0, maximum: 359, nullable: true }
        palette:
          type: array
          items: { type: string }
        sceneTags:
          type: array
          items: { type: string }
        objectTags:
          type: array
          items: { type: string }
        motionScore: { type: number, format: float, nullable: true }
        favorite: { type: boolean, default: false }
        altText: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
...